requirements/requirements.txt
asgiref==3.6.0
boto3==1.26.35
botocore==1.29.35
Django==4.1.4
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-storages==1.13.1
ImageHash==4.3.1
jmespath==1.0.1
numpy==1.24.0
Pillow==9.3.0
psycopg2==2.9.5
python-dateutil==2.8.2
PyWavelets==1.4.1
s3transfer==0.6.0
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
